feat: redesign the profile generator to be a CP-SAT solver - refactored how constraints are stored and verified

+++ backend/types/ceab_attributes.py
@@ -3,22 +3,22 @@ class CEABAttributes:
     """Represents CEAB accreditation attribute breakdown for a course."""
 
     def __init__(self,
-                 total_AU: int = 0,
-                 mathematics: int = 0,
-                 natural_science: int = 0,
-                 math_and_science: int = 0,
-                 engineering_science: int = 0,
-                 engineering_design: int = 0,
-                 eng_sci_and_design: int = 0,
-                 complementary_studies: int = 0):
-        self.total_AU = total_AU
-        self.mathematics = mathematics
-        self.natural_science = natural_science
-        self.math_and_science = math_and_science
-        self.engineering_science = engineering_science
-        self.engineering_design = engineering_design
-        self.eng_sci_and_design = eng_sci_and_design
-        self.complementary_studies = complementary_studies
+                 total_AU: float = 0.0,
+                 mathematics: float = 0.0,
+                 natural_science: float = 0.0,
+                 math_and_science: float = 0.0,
+                 engineering_science: float = 0.0,
+                 engineering_design: float = 0.0,
+                 eng_sci_and_design: float = 0.0,
+                 complementary_studies: float = 0.0):
+        self.total_AU = float(total_AU)
+        self.mathematics = float(mathematics)
+        self.natural_science = float(natural_science)
+        self.math_and_science = float(math_and_science)
+        self.engineering_science = float(engineering_science)
+        self.engineering_design = float(engineering_design)
+        self.eng_sci_and_design = float(eng_sci_and_design)
+        self.complementary_studies = float(complementary_studies)
 
     def __repr__(self):
         return (f"CEABAttributes(TotalAU={self.total_AU}, Math={self.mathematics}, "
